Reorganizes initialization process in EmStencil
- Refactored `initialize.py` to improve organization and clarify logic.
- Separated initialization logic into functions for better modularity.
- Updated `__init__.py` to use the new initialization functions.

This change improves the readability and maintainability of the initialization code, making it easier to manage and extend in the future.
